backend-py/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
aiohttp>=3.8.5
pydantic>=2.4.2
